requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==5.0.1
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
barcodenumber==0.5.0
beautifulsoup4==4.12.3
bleach==6.0.0
bleach-allowlist==1.0.3
boto3==1.34.162
botocore==1.34.162
braintree==4.20.0
Brotli==1.1.0
cachetools==4.2.4
cairocffi==1.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.0
click==8.1.7
croniter==2.0.7
cryptography==43.0.3
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.9.0
dataclasses-json==0.5.14
decorator==5.1.1
-e git+https://github.com/aliraxa29/desk_pos_frappe.git@d3cf76bf734a83e7e5e90abc5b4e4f8aa140f54e#egg=desk_pos
docopt==0.6.2
dropbox==11.36.2
email-reply-parser==0.5.12
-e git+https://github.com/frappe/erpnext.git@81f1f1f1bbc7c8947776981873dd8b8794aa4d1b#egg=erpnext
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.4
filetype==1.2.0
fonttools==4.55.0
-e git+https://github.com/frappe/frappe.git@ab7e080e0c39fecbdcd16e97c6375e087a079844#egg=frappe
gitdb==4.0.11
GitPython==3.1.43
gocardless-pro==1.22.0
google-api-core==1.34.1
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.66.0
googlemaps==4.10.0
gunicorn==22.0.0
hiredis==2.2.3
holidays==0.61
-e git+https://github.com/frappe/hrms.git@6f7a7b586fb18772f0a23ca847a1650d26856992#egg=hrms
html5lib==1.1
httplib2==0.22.0
idna==3.10
ipython==8.15.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
ldap3==2.9.1
lxml==5.3.0
markdown2==2.4.13
markdownify==0.11.6
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib-inline==0.1.7
maxminddb==2.6.2
maxminddb-geolite2==2018.703
monotonic==1.6
mypy-extensions==1.0.0
num2words==0.5.13
oauthlib==3.2.2
-e git+https://github.com/aliraxa29/ocr.git@c4f66075d2be4eac90825da7966592b261075205#egg=ocr
openpyxl==3.1.5
packaging==24.2
parso==0.8.4
passlib==1.7.4
-e git+https://github.com/frappe/payments.git@b5eda73300770c7d0e07f31e5011772adbc0ceca#egg=payments
paytmchecksum==1.7.0
pdfkit==1.0.0
pexpect==4.9.0
phonenumbers==8.13.13
pickleshare==0.7.5
pillow==10.3.0
plaid-python==7.2.1
ply==3.11
posthog==3.0.2
premailer==3.10.0
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==22.3.5
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.7.4
pydantic_core==2.18.4
pydyf==0.10.0
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyotp==2.8.0
pyparsing==3.2.0
pypdf==3.17.4
pyphen==0.17.0
PyPika==0.48.9
pypng==0.20220715.0
PyQRCode==1.2.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-stdnum==1.20
python-youtube==0.8.3
pytz==2023.3
PyYAML==6.0.2
rapidfuzz==2.15.2
rauth==0.7.3
razorpay==1.2.0
redis==4.5.5
requests==2.32.3
requests-oauthlib==1.3.1
RestrictedPython==7.4
rq==1.15.1
rsa==4.9
s3transfer==0.10.4
semantic-version==2.10.0
sentry-sdk==1.37.1
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sql_metadata==2.11.0
sqlparse==0.5.2
stack-data==0.6.3
stone==3.3.8
stripe==2.56.0
tenacity==8.2.3
terminaltables==3.1.10
tinycss2==1.1.1
tomli==2.0.2
traceback-with-variables==2.0.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
Unidecode==1.3.8
uritemplate==3.0.1
urllib3==2.2.3
wcwidth==0.2.13
weasyprint==59.0
webencodings==0.5.1
Werkzeug==3.0.6
Whoosh==2.7.4
-e git+https://github.com/xconsultingAi/xhr.git@704695d2f738822cdb22dfe09a5d891af1d7ffba#egg=xhr
xlrd==2.0.1
zopfli==0.2.3.post1
zxcvbn==4.4.28
